# feat(skipcpio): speed up and harden skipcpio