@haraldh 2021-02-25
> Add the missing libcpio copyright as suggested by @ddiss

@johannbg removed the copied code